navierflow/core/compute_engine.py
import torch
import triton
import triton.language as tl
import numpy as np
from typing import Dict, List, Optional, Tuple
from enum import Enum
import logging

class ComputeBackend(Enum):
    CUDA = "cuda"
    TRITON = "triton"
    HYBRID = "hybrid"

@triton.jit
def linear_solver_kernel(
    ptr_a, ptr_b, ptr_x,
    stride_am, stride_ak, stride_bk,
    M, K,
    BLOCK_SIZE: tl.constexpr
):
    """Triton kernel for linear system solver"""
    pid = tl.program_id(0)
    block_start = pid * BLOCK_SIZE
    
    # Load matrix A and vector b
    offs_am = block_start + tl.arange(0, BLOCK_SIZE)
    offs_ak = tl.arange(0, BLOCK_SIZE)
    a = tl.load(ptr_a + offs_am[:, None] * stride_am + offs_ak[None, :] * stride_ak)
    b = tl.load(ptr_b + offs_am * stride_bk)
    
    # Solve using parallel reduction
    x = tl.sum(a * b[:, None], axis=1)
    
    # Store result
    tl.store(ptr_x + offs_am, x)

@triton.jit
def pressure_poisson_kernel(
    ptr_pressure, ptr_divergence,
    stride_ph, stride_pw,
    H, W,
    BLOCK_SIZE: tl.constexpr
):
    """Triton kernel for pressure Poisson equation"""
    pid = tl.program_id(0)
    block_start = pid * BLOCK_SIZE
    
    # Load pressure and divergence
    offs_h = block_start + tl.arange(0, BLOCK_SIZE)
    offs_w = tl.arange(0, BLOCK_SIZE)
    p = tl.load(ptr_pressure + offs_h[:, None] * stride_ph + offs_w[None, :] * stride_pw)
    div = tl.load(ptr_divergence + offs_h[:, None] * stride_ph + offs_w[None, :] * stride_pw)
    
    # Compute Laplacian
    laplacian = (
        tl.load(ptr_pressure + (offs_h[:, None] + 1) * stride_ph + offs_w[None, :] * stride_pw) +
        tl.load(ptr_pressure + (offs_h[:, None] - 1) * stride_ph + offs_w[None, :] * stride_pw) +
        tl.load(ptr_pressure + offs_h[:, None] * stride_ph + (offs_w[None, :] + 1) * stride_pw) +
        tl.load(ptr_pressure + offs_h[:, None] * stride_ph + (offs_w[None, :] - 1) * stride_pw) -
        4.0 * p
    )
    
    # Update pressure
    p_new = p + 0.25 * (laplacian - div)
    tl.store(ptr_pressure + offs_h[:, None] * stride_ph + offs_w[None, :] * stride_pw, p_new)

class OptimizedComputeEngine:
    def __init__(self, backend: ComputeBackend = ComputeBackend.HYBRID):
        self.backend = backend
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        
        # Initialize CUDA handles
        if backend in [ComputeBackend.CUDA, ComputeBackend.HYBRID]:
            try:
                import cupy as cp
                self.cublas_handle = cp.cuda.cublas.create()
                self.cusparse_handle = cp.cuda.cusparse.create()
            except ImportError:
                logging.warning("CuPy not available. Falling back to PyTorch.")
        
        # Performance metrics
        self.metrics = {
            'linear_solve_time': 0.0,
            'poisson_solve_time': 0.0,
            'memory_usage': 0.0
        }

    def solve_linear_system(self, A: torch.Tensor, b: torch.Tensor,
                          method: str = 'direct') -> torch.Tensor:
        """Solve linear system Ax = b using optimized backends"""
        if method == 'direct':
            return self._solve_direct(A, b)
        elif method == 'iterative':
            return self._solve_iterative(A, b)
        else:
            raise ValueError(f"Unknown solver method: {method}")

    def _solve_direct(self, A: torch.Tensor, b: torch.Tensor) -> torch.Tensor:
        """Direct solver using cuBLAS/cuSPARSE"""
        start_event = torch.cuda.Event(enable_timing=True)
        end_event = torch.cuda.Event(enable_timing=True)
        
        start_event.record()
        
        if self.backend in [ComputeBackend.CUDA, ComputeBackend.HYBRID]:
            try:
                import cupy as cp
                A_gpu = cp.asarray(A.detach().cpu().numpy())
                b_gpu = cp.asarray(b.detach().cpu().numpy())
                
                # Use cuBLAS for dense solver
                x_gpu = cp.linalg.solve(A_gpu, b_gpu)
                x = torch.from_numpy(cp.asnumpy(x_gpu)).to(self.device)
            except:
                # Fallback to PyTorch
                x = torch.linalg.solve(A, b)
        else:
            # Use Triton kernel for custom implementation
            M, K = A.shape
            x = torch.zeros((M,), device=self.device)
            grid = lambda meta: (triton.cdiv(M, meta['BLOCK_SIZE']),)
            
            linear_solver_kernel[grid](
                A, b, x,
                A.stride(0), A.stride(1), b.stride(0),
                M, K,
                BLOCK_SIZE=32
            )
        
        end_event.record()
        end_event.synchronize()
        self.metrics['linear_solve_time'] = start_event.elapsed_time(end_event)
        
        return x

    def _solve_iterative(self, A: torch.Tensor, b: torch.Tensor,
                        max_iter: int = 1000, tol: float = 1e-6) -> torch.Tensor:
        """Iterative solver using conjugate gradient"""
        x = torch.zeros_like(b)
        r = b - torch.matmul(A, x)
        p = r.clone()
        r_norm = torch.dot(r, r)
        
        for i in range(max_iter):
            Ap = torch.matmul(A, p)
            alpha = r_norm / torch.dot(p, Ap)
            x += alpha * p
            r -= alpha * Ap
            r_norm_new = torch.dot(r, r)
            beta = r_norm_new / r_norm
            r_norm = r_norm_new
            if r_norm < tol:
                break
            p = r + beta * p
        
        return x

    def solve_pressure_poisson(self, pressure: torch.Tensor,
                             divergence: torch.Tensor,
                             num_iterations: int = 50) -> torch.Tensor:
        """Solve pressure Poisson equation using optimized kernels"""
        start_event = torch.cuda.Event(enable_timing=True)
        end_event = torch.cuda.Event(enable_timing=True)
        
        start_event.record()
        
        H, W = pressure.shape
        grid = lambda meta: (triton.cdiv(H, meta['BLOCK_SIZE']),)
        
        for _ in range(num_iterations):
            pressure_poisson_kernel[grid](
                pressure, divergence,
                pressure.stride(0), pressure.stride(1),
                H, W,
                BLOCK_SIZE=32
            )
        
        end_event.record()
        end_event.synchronize()
        self.metrics['poisson_solve_time'] = start_event.elapsed_time(end_event)
        
        return pressure

    def compute_gradients(self, field: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
        """Compute gradients using optimized finite difference"""
        dx = torch.zeros_like(field)
        dy = torch.zeros_like(field)
        
        # Central differences for interior
        dx[1:-1, 1:-1] = (field[1:-1, 2:] - field[1:-1, :-2]) / 2.0
        dy[1:-1, 1:-1] = (field[2:, 1:-1] - field[:-2, 1:-1]) / 2.0
        
        # Forward/backward differences for boundaries
        dx[1:-1, 0] = field[1:-1, 1] - field[1:-1, 0]
        dx[1:-1, -1] = field[1:-1, -1] - field[1:-1, -2]
        dy[0, 1:-1] = field[1, 1:-1] - field[0, 1:-1]
        dy[-1, 1:-1] = field[-1, 1:-1] - field[-2, 1:-1]
        
        return dx, dy

    def get_performance_metrics(self) -> Dict:
        """Get performance metrics"""
        # Update memory usage
        self.metrics['memory_usage'] = torch.cuda.max_memory_allocated() / 1e9  # GB
        return self.metrics.copy()

    def __del__(self):
        """Cleanup CUDA handles"""
        if hasattr(self, 'cublas_handle'):
            self.cublas_handle.destroy()
        if hasattr(self, 'cusparse_handle'):
            self.cusparse_handle.destroy() 