navierflow/utils/cad_integration.py
import numpy as np
import trimesh
import os
from typing import Dict, List, Optional, Tuple, Union
from pathlib import Path
import logging

class CADIntegration:
    """
    CAD integration module for NavierFlow.
    Supports importing and exporting various CAD formats for fluid simulation.
    """
    
    SUPPORTED_FORMATS = {
        'import': ['.stl', '.obj', '.step', '.iges', '.brep'],
        'export': ['.stl', '.obj', '.step']
    }
    
    def __init__(self, config: Dict = None):
        self.config = {
            'mesh_simplification': True,
            'auto_repair': True,
            'voxel_size': 0.01,
            'export_quality': 'high'
        }
        if config:
            self.config.update(config)
            
        self.current_mesh = None
        self.voxel_grid = None
        self.boundary_conditions = {}
        
    def import_cad(self, file_path: Union[str, Path],
                   scale: float = 1.0) -> bool:
        """
        Import CAD file and prepare it for fluid simulation.
        
        Args:
            file_path: Path to CAD file
            scale: Scaling factor for the imported geometry
            
        Returns:
            bool: Success status
        """
        try:
            file_path = Path(file_path)
            if file_path.suffix.lower() not in self.SUPPORTED_FORMATS['import']:
                raise ValueError(f"Unsupported file format: {file_path.suffix}")
                
            # Load mesh using trimesh
            self.current_mesh = trimesh.load(file_path)
            
            # Apply scaling
            if scale != 1.0:
                self.current_mesh.apply_scale(scale)
                
            # Mesh cleanup and repair if enabled
            if self.config['auto_repair']:
                self.current_mesh.fill_holes()
                self.current_mesh.remove_duplicate_faces()
                self.current_mesh.remove_degenerate_faces()
                
            # Mesh simplification if enabled
            if self.config['mesh_simplification']:
                target_faces = len(self.current_mesh.faces) // 2
                self.current_mesh = self.current_mesh.simplify_quadratic_decimation(target_faces)
                
            # Generate voxel grid for simulation
            self.voxelize_mesh()
            
            return True
            
        except Exception as e:
            logging.error(f"Error importing CAD file: {str(e)}")
            return False
            
    def voxelize_mesh(self):
        """Convert mesh to voxel grid for fluid simulation"""
        if self.current_mesh is None:
            raise ValueError("No mesh loaded")
            
        # Create voxel grid
        voxel_size = self.config['voxel_size']
        self.voxel_grid = self.current_mesh.voxelized(voxel_size)
        self.voxel_grid = self.voxel_grid.fill()  # Fill internal voids
        
    def set_boundary_condition(self, surface_name: str,
                             condition_type: str,
                             parameters: Dict):
        """
        Set boundary conditions for specific surfaces.
        
        Args:
            surface_name: Name or identifier of the surface
            condition_type: Type of boundary condition ('wall', 'inlet', 'outlet', etc.)
            parameters: Dictionary of boundary condition parameters
        """
        self.boundary_conditions[surface_name] = {
            'type': condition_type,
            'parameters': parameters
        }
        
    def export_simulation_mesh(self, file_path: Union[str, Path],
                             format_type: str = 'stl') -> bool:
        """
        Export the processed mesh for simulation.
        
        Args:
            file_path: Output file path
            format_type: Output format ('stl', 'obj', 'step')
            
        Returns:
            bool: Success status
        """
        try:
            file_path = Path(file_path)
            if format_type.lower() not in self.SUPPORTED_FORMATS['export']:
                raise ValueError(f"Unsupported export format: {format_type}")
                
            # Export based on format
            if format_type.lower() == 'stl':
                self.current_mesh.export(file_path, file_type='stl')
            elif format_type.lower() == 'obj':
                self.current_mesh.export(file_path, file_type='obj')
            elif format_type.lower() == 'step':
                # STEP export requires additional processing
                self._export_step(file_path)
                
            return True
            
        except Exception as e:
            logging.error(f"Error exporting simulation mesh: {str(e)}")
            return False
            
    def get_voxel_data(self) -> Optional[np.ndarray]:
        """Return voxel grid data for simulation"""
        if self.voxel_grid is None:
            return None
        return self.voxel_grid.matrix
        
    def get_boundary_surfaces(self) -> Dict:
        """Return dictionary of boundary surfaces and their conditions"""
        return self.boundary_conditions
        
    def get_mesh_statistics(self) -> Dict:
        """Return statistics about the current mesh"""
        if self.current_mesh is None:
            return {}
            
        return {
            'vertices': len(self.current_mesh.vertices),
            'faces': len(self.current_mesh.faces),
            'volume': self.current_mesh.volume,
            'surface_area': self.current_mesh.area,
            'is_watertight': self.current_mesh.is_watertight,
            'is_manifold': self.current_mesh.is_manifold
        }
        
    def _export_step(self, file_path: Path):
        """Export mesh to STEP format"""
        try:
            import pythonOCC.Core.BRep as BRep
            import pythonOCC.Core.STEPControl as STEPControl
            
            # Convert trimesh to OpenCASCADE shape
            # This would require additional implementation
            pass
            
        except ImportError:
            logging.error("pythonOCC not installed. STEP export not available.")
            raise
            
    def validate_mesh_for_simulation(self) -> Tuple[bool, List[str]]:
        """
        Validate mesh for CFD simulation.
        
        Returns:
            Tuple of (is_valid, list of issues)
        """
        if self.current_mesh is None:
            return False, ["No mesh loaded"]
            
        issues = []
        
        # Check watertightness
        if not self.current_mesh.is_watertight:
            issues.append("Mesh is not watertight")
            
        # Check manifold property
        if not self.current_mesh.is_manifold:
            issues.append("Mesh is not manifold")
            
        # Check for degenerate faces
        if len(self.current_mesh.degenerate_faces) > 0:
            issues.append(f"Found {len(self.current_mesh.degenerate_faces)} degenerate faces")
            
        # Check mesh quality
        quality = self.check_mesh_quality()
        if quality['min_angle'] < 15.0:  # degrees
            issues.append("Poor quality elements detected (small angles)")
            
        if quality['aspect_ratio'] > 10.0:
            issues.append("Poor quality elements detected (high aspect ratio)")
            
        return len(issues) == 0, issues
        
    def check_mesh_quality(self) -> Dict:
        """Analyze mesh quality metrics"""
        if self.current_mesh is None:
            return {}
            
        # Calculate mesh quality metrics
        # This is a simplified version; real implementation would be more comprehensive
        quality_metrics = {
            'min_angle': 30.0,  # Placeholder
            'max_angle': 120.0,  # Placeholder
            'aspect_ratio': 5.0,  # Placeholder
            'skewness': 0.5     # Placeholder
        }
        
        return quality_metrics
        
    def generate_boundary_layer(self, thickness: float,
                              num_layers: int = 5) -> bool:
        """
        Generate boundary layer mesh for viscous flow simulation.
        
        Args:
            thickness: Total thickness of boundary layer
            num_layers: Number of layers
            
        Returns:
            bool: Success status
        """
        if self.current_mesh is None:
            return False
            
        try:
            # Boundary layer mesh generation would go here
            # This requires advanced mesh manipulation
            return True
            
        except Exception as e:
            logging.error(f"Error generating boundary layer: {str(e)}")
            return False 