requirements.txt
# Core dependencies
taichi>=1.7.0
numpy>=1.21.0
torch>=2.1.0
trimesh>=3.9.0
pythonOCC-core>=7.4.0

# Scientific computing
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
plotly>=5.18.0
pyvista>=0.34.0

# Machine learning
scikit-learn>=1.0.0
tensorboard>=2.7.0

# CAD and mesh processing
meshio>=5.0.0
gmsh>=4.8.0

# Utilities
tqdm>=4.62.0
PyYAML>=5.4.0
rich>=10.12.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0

# Development tools
black>=21.9b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910

# AI and Machine Learning
wandb>=0.16.0
h5py>=3.8.0

# Visualization and GUI
seaborn>=0.12.0

# Data processing and utilities
pyyaml>=6.0.0
tqdm>=4.65.0

# Development tools
pytest>=7.0.0
mypy>=1.0.0
flake8>=6.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
nbsphinx>=0.9.0

# Optional: CUDA support for PyTorch
# Install appropriate version based on your CUDA version
# torch-cuda

# Physics and Numerics
petsc4py>=3.17.0
slepc4py>=3.17.0

# Visualization and UI
streamlit>=1.28.0
opencv-python>=4.7.0
pillow>=9.4.0
dearpygui>=1.9.0
ipywidgets>=8.0.0
bokeh>=3.1.0

# Data Processing and Analysis
xarray>=2023.1.0

# Performance Monitoring
psutil>=5.9.0
py-cpuinfo>=9.0.0
gputil>=1.4.0

# Export and Conversion
ffmpeg-python>=0.2.0
imageio>=2.25.0
moviepy>=1.0.3

# Development Tools
pre-commit>=3.2.0

# Documentation
jupyter>=1.0.0

# Optional: CUDA Support
cupy>=12.0.0
pycuda>=2022.2.2

# AI/ML
tensorflow>=2.12.0
pytorch-lightning>=2.0.0
optuna>=3.1.0
mlflow>=2.3.0

# Web Framework
fastapi>=0.95.0
uvicorn>=0.21.0
websockets>=11.0.0
httpx>=0.24.0
pydantic>=1.10.0
jwt>=1.3.1

# Cloud and HPC
kubernetes>=26.1.0
docker>=6.0.0
mpi4py>=3.1.0
dask>=2023.3.0
ray>=2.3.0
horovod>=0.27.0

# CAD Integration
pythreejs>=2.4.0
cadquery>=2.1
freecad>=0.20.0

# Development and Testing
pytest-benchmark>=4.0.0

# Monitoring and Profiling
prometheus-client>=0.16.0
grafana-client>=2.1.0
pyinstrument>=4.4.0
memory-profiler>=0.61.0

# Logging and Monitoring
sentry-sdk>=1.19.0
elastic-apm>=6.15.0

# Version Control
gitpython>=3.1.0 